Change tuple-variant expansion to use rest syntax.
So a little birdie in the Rust Community Discord Server told me that
tuple patterns, such as those in tuple structs and variants, support use
of the .. ("rest") operator, instead of having to manually ignore every
position. This greatly simplifies the output assembly code for these
variants. Thanks!

Also turned on link-time optimisation because *why not.*